+++ server/src/main/java/com/ifmo/adda/service/OrdersService.kt
@@ -19,7 +19,7 @@ class OrdersService(
     )
 
     @Throws(Exception::class)
-    fun getOrdersForClient(clientId: BigInteger):OrdersList {
+    fun getOrdersForClient(clientId: Int):OrdersList {
         val orders = ordersRepository.findNormalForClient(clientId).map { it.toDto() } +
                 customOrdersRepository.findCustomForClient(clientId).map { it.toDto() }
         if (orders.isEmpty()) throw Exception("Orders for client with id $clientId not found")
